begin endMeeting unit testing
